Split helper so it can be used in edge environments
