+++ packages/next/src/server/next-server.ts
@@ -91,7 +91,8 @@ import { getTracer } from './lib/trace/tracer'
 import { NextNodeServerSpan } from './lib/trace/constants'
 import { nodeFs } from './lib/node-fs-methods'
 import { getRouteRegex } from '../shared/lib/router/utils/route-regex'
-import { invokeRequest, pipeReadable } from './lib/server-ipc/invoke-request'
+import { invokeRequest } from './lib/server-ipc/invoke-request'
+import { pipeReadable } from './pipe-readable'
 import { filterReqHeaders } from './lib/server-ipc/utils'
 import { createRequestResponseMocks } from './lib/mock-request'
 import chalk from 'next/dist/compiled/chalk'
